refactor: Move `Wizard` into a separate file This also includes a drive-by fix to remove an unused state variable.
