+++ macos/Fileaway/Scenes/Wizard.swift
@@ -1,0 +1,46 @@
+// Copyright (c) 2018-2022 InSeven Limited
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+import SwiftUI
+
+struct Wizard: Scene {
+
+    static let windowID = "wizard-window"
+
+    private var manager: Manager
+
+    init(manager: Manager) {
+        self.manager = manager
+    }
+
+    var body: some Scene {
+        WindowGroup(id: Self.windowID, for: URL.self) { $url in
+            if let url = url {
+                RulesWizard(url: url)
+                    .environment(\.manager, manager)
+                    .background(VisualEffectView(material: NSVisualEffectView.Material.sidebar,
+                                                 blendingMode: NSVisualEffectView.BlendingMode.behindWindow)
+                        .edgesIgnoringSafeArea(.all))
+            }
+        }
+        .windowStyle(.hiddenTitleBar)
+    }
+
+}
